+++ etils/epy/__init__.py
@@ -19,6 +19,7 @@
 import sys
 
 from etils.epy.backports import cached_property
+from etils.epy.binary_import import binary_adhoc
 from etils.epy.contextlib import ContextManager
 from etils.epy.env_utils import is_ipython_terminal
 from etils.epy.env_utils import is_notebook