+++ CHANGELOG.md
@@ -11,7 +11,8 @@ Changelog follow https://keepachangelog.com/ format.
 *   `ecolab`:
     *   Added protobuf repeated fields support to `ecolab.inspect`
     *   `ecolab.auto_display`:
-        *   Support specifiers to customize auto-display (`;s`, `;a`, `;i`,...)
+        *   Support specifiers to customize auto-display (`;s`, `;a`, `;i`,
+            `;p`,...)
         *   Fixed auto-display when the line contain UTF-8 character
 
 ## [1.5.2] - 2023-10-24