+++ CHANGELOG.md
@@ -20,6 +20,9 @@ Changelog follow https://keepachangelog.com/ format.
         `ecolab.adhoc` context when resolved)
     *   Added: `epy.is_ipython_terminal()` to detect when running in
         non-notebook terminal environment.
+    *   Added: `epy.pretty_repr` support `attr` dataclass-like objects.
+*   `exm`:
+    *   Adding artifacts can be used inside `with xm.create_experiment()`
 
 ## [1.5.2] - 2023-10-24